// src/types/shift.ts
export type ShiftKasir = {
  shift_date: string;
  id: string;
  user_id: string;
  cashier_name: string;
  opened_at: string;
  closed_at: string;
  opening_balance: number;
  closing_balance: number;
  system_balance: number;
  difference: number;
  closing_status: string;
  closed_by: string;
};